advisor_logging: Extracted logging features
